Moved all debug UI elements into a new DebugMenu class
